Add files via upload
Client data such as: age of the client, their job type, their marital status, etc. Along with the client data, the information of the call such as the duration of the call, day and month of the call, etc is provided. This data is fed into a predictive model to predict whether the customer will subscribe to term deposit.